Fix typo?
Like Carolyn, I was confused by the two ones indicating three bytes, given the previous slide says "the number of 1's tells UTF-8 how many bytes there are". It doesn't seem like it can be a binary representation of an integer, because the first 0 denotes the end of the header. I've messed with unicode a lot, but must confess never at the bit level.