Use no style at all for .out inclusions.
The Remark slideshow framework uses Highlight.js to style code.
Unfortunately it doesn't have a style for '.out', and its style
for other suffixes we might use for program output (like '.txt')
doesn't render in slides. The simplest thing for now is therefore
to strip off the format entirely for `.out` inclusions.

Closes #33.