Fixed bug in VM debugger for adding/clearing breakpoints
If you try adding a breakpoint and this operation fails (because there is already a breakpoint at that address), the debugger will step to the next instruction, which seems to be unintended behaviour. The same happens for trying to clear a breakpoint at an address where there is no breakpoint. This happens because these two commands return None in case of failure, but that None interferes with the loop from `interact` in `vm_step.py`